# Improve performance of MatrixFourierTransform