--- a/hcipy/config/config.py
+++ b/hcipy/config/config.py
@@ -1,12 +1,63 @@
 import os
 import yaml
+import pkg_resources
+
+class _ConfigurationItem(object):
+	def __init__(self, val):
+		self._val = val
+
+	def __getitem__(self, key):
+		val = self._val[key]
+
+		if isinstance(val, dict):
+			return _ConfigurationItem(val)
+		else:
+			return val
+
+	def __setitem__(self, key, value):
+		self._val[key] = value
+
+	def __contains__(self, key):
+		return key in self._val
+
+	def __getattr__(self, name):
+		if name == '_val':
+			return getattr(self, name)
+		else:
+			return self.__getitem__(name)
+
+	def __setattr__(self, name, value):
+		if name == '_val':
+			super().__setattr__(name, value)
+		else:
+			self.__setitem__(name, value)
+
+	def __str__(self):
+		return str(self._val)
+
+	def clear(self):
+		self._val.clear()
+
+	def update(self, b):
+		for key in b.keys():
+			if key in self:
+				a = self[key]
+
+				if isinstance(b[key], dict) and isinstance(self._val[key], dict):
+					self[key].update(b[key])
+				else:
+					self[key] = b[key]
+			else:
+				self[key] = b[key]
 
 class Configuration(object):
 	'''A configuration object that describes the current configuration status of the package.
 	'''
 	def __init__(self):
-		if not hasattr(Configuration, '_config'):
+		if Configuration._config is None:
 			self.reset()
+
+	_config = None
 
 	def __getitem__(self, key):
 		'''Get the value for the key.
@@ -30,53 +81,46 @@ class Configuration(object):
 		'''
 		Configuration._config[key] = value
 
-	def get_config_path(self):
-		'''Get the current path of the configuration file.
+	__getattr__ = __getitem__
+	__setattr__ = __setitem__
 
-		If no configuration file is present, a new one will be made,
-		and its path will be returned.
-
-		Returns
-		-------
-		string
-			The configuration directory.
-		'''
-		paths = ['./']
-		if 'XDG_CONFIG_HOME' in os.environ:
-			for p in os.environ['XDG_CONFIG_HOME'].split(os.pathsep):
-				paths.append(p + '/hcipy/')
-		if 'XDG_CONFIG_DIRS' in os.environ:
-			for p in os.environ['XDG_CONFIG_DIRS'].split(os.pathsep):
-				paths.append(p + '/hcipy/')
-		paths.append(os.path.expanduser('~/.config/hcipy/'))
-		paths.append(os.path.expanduser('~/.hcipy/'))
-		paths = [p + 'hcipy_config.yaml' for p in paths]
-
-		for path in paths:
-			if os.path.exists(path):
-				return path
-
-		# No configuration file is available, make a new one.
-		path = paths[-1]
-		directory = os.path.dirname(path)
-		if not os.path.exists(directory):
-			os.makedirs(directory)
-
-		f = open(path, 'w')
-		f.write(default_configuration)
-		f.close()
-
-		return path
+	def __str__(self):
+		return str(Configuration._config)
 
 	def reset(self):
-		'''Read the configuration file from the configuration directory.
+		'''Reset the configuration to the default configuration.
+
+		This default configuration consists of the default parameters in `hcipy/data/default_config.yaml`, which
+		can be overridden by a configuration file in `~/.hcipy/hcipy_config.yaml`. This can in turn be overridden
+		by a configuration file named `hcipy_config.yaml` located in the current working directory.
 		'''
-		with open(self.get_config_path()) as f:
-			Configuration._config = yaml.safe_load(f.read())
+		Configuration._config = _ConfigurationItem({})
 
-		if Configuration._config is None:
-			Configuration._config = {}
+		default_config = pkg_resources.resource_stream('hcipy', 'data/default_config.yaml')
+		user_config = os.path.expanduser('~/.hcipy/hcipy_config.yaml')
+		current_working_directory = './hcipy_config.yaml'
 
-default_configuration = '''# This is the configuration file for the high-contrast imaging for Python (HCIPy) library.
+		paths = [default_config, user_config, current_working_directory]
 
-'''
+		for path in paths:
+			try:
+				contents = path.read()
+			except AttributeError:
+				try:
+					with open(path) as f:
+						contents = f.read()
+				except IOError:
+					continue
+
+			new_config = yaml.safe_load(contents)
+			self.update(new_config)
+
+	def update(self, b):
+		'''Update the configuration with the configuration `b`, as a dictionary.
+
+		Parameters
+		---
+		b : dict
+			A dictionary containing the values to update in the configuration.
+		'''
+		Configuration._config.update(b)