Overview
===
I replaced the np.dot()s with manual calls to BLAS-GEMM, which allow for taking the adjoint/transpose of the input matrices for free. The latter is used for the backward transform, allowing it to not calculate the adjoint beforehand with numpy. I also optimized the plane in which the weights are applied: in case the weights are the same for every pixel, you can apply the weight either the input or output plane, and we choose the one which is quickest (ie. smallest number of pixels).

The overhead induced by HCIPy before was ~10-50%, depending on the size of the arrays. This is now <1%.

The direct GEMM call is only applied for complex128, as the matrices of the MFT are stored using that data type. For other data types, the usual np.dot() calls are used.

Benchmarking results
===

Performed on a Intel Core i7-9750H @ 2.6GHz. Speed improvement on forward: +9% to +186%, depending on array sizes. Speed improvement on backward: +32% to +77%, depending on array sizes.

256x256->64x64
---
Forward old: 1.13 ms ± 38.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Backward old: 1.85 ms ± 129 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Forward new: 395 µs ± 31.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Backward new: 1.07 ms ± 29.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Improvement: +186% (forward) and +73% (backward)

512x512->256x256
---
Forward old: 11.9 ms ± 999 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Backward old: 17.7 ms ± 1.23 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
Forward new: 8.46 ms ± 561 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Backward new: 11.8 ms ± 198 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Improvement: +41% (forward) and +50% (backward)

2048x2048->256x256
---
Forward old: 127 ms ± 6.98 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward old: 155 ms ± 3.8 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Forward new: 71.6 ms ± 1.67 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward new: 117 ms ± 5.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Improvement: +77% (forward) and +32% (backward)

256x256->256x256
---
Forward old: 4.32 ms ± 333 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Backward old: 6.65 ms ± 474 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Forward new: 3.44 ms ± 354 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Backward new: 3.76 ms ± 568 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Improvement: +26% (forward) and +77% (backward)

1024x1024->1024x1024
---
Forward old: 146 ms ± 2.56 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward old: 203 ms ± 9.22 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Forward new: 134 ms ± 1.43 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward new: 141 ms ± 3.06 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Improvement: +9% (forward) and +43% (backward)