@ehpor 2020-05-13
I did the same for the FFT. I implemented the (i)FFTshifts as multiplications in the opposite domain. This removes the FFTshifts but decreases the accuracy. This is why I introduced the variable `high_accuracy`, which can be turned on if high accuracy is required (at the cost of runtime). The loss in accuracy is minor (about a factor of 10), which shouldn't be noticable in 64bit mode (error is still <1e-13, so <1e-26 in intensity, for most fields).

Speed improvement is ~3-5x on both forward and backward transforms.

Benchmarking results
===

Performed on a Intel Core i7-9750H @ 2.6GHz.

512x512, q=1, fov=1
---
Forward old: 20.7 ms ± 1.11 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward old: 23.9 ms ± 1.26 ms per loop (mean ± std. dev. of 7 runs, 10 loops each) 
Forward new: 4.08 ms ± 158 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Backward new: 4.43 ms ± 192 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Improvement: 5.1x (forward), 5.4x (backward) 

512x512, q=4, fov=0.125
---
Forward old: 162 ms ± 2.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward old: 157 ms ± 5.78 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Forward new: 42 ms ± 2.16 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward new: 44.6 ms ± 1.04 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Improvement: 3.9x (forward), 3.5x (backward)

2048x2048, q=1, fov=1
---
Forward old: 277 ms ± 10.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Backward old: 321 ms ± 30.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) 
Forward new: 79 ms ± 1.43 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Backward new: 82.8 ms ± 1.33 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Improvement: 3.5x (forward), 3.9x (backward)